refactoring and fix #38
