+++ src/CronExpression.php
@@ -13,21 +13,18 @@ class CronExpression
     public DaysOfMonthField $day;
     public MonthsField $month;
     public DaysOfWeekField $weekday;
-    public string $locale;
     public bool $timeFormat24hours;
     public array $translations;
 
+
     /**
-     * Constructor
-     * 
-     * @param string $cron The cron expression
-     * @param string $locale The locale 
-     * @param bool $timeFormat24hours Whether to use 24 hour time
+     * Constructs a new instance of the class.
      *
-     * @throws CronParsingException
-     * @throws TranslationFileMissingException  
+     * @param string $cron The cron expression.
+     * @param array $translations Optional translations for field values.
+     * @param bool $timeFormat24hours Whether to use 24-hour time format.
      */
-    public function __construct(string $cron, string $locale = 'en', bool $timeFormat24hours = false)
+    public function __construct(string $cron, array $translations = [], bool $timeFormat24hours = false)
     {
         $this->raw = $cron;
         $fields = explode(' ', $cron);
@@ -38,11 +35,9 @@ class CronExpression
         $this->month = new MonthsField($this, $fields[3]);
         $this->weekday = new DaysOfWeekField($this, $fields[4]);
 
-        $this->locale = $locale;
         $this->timeFormat24hours = $timeFormat24hours;
 
-        $this->ensureLocaleExists();
-        $this->loadTranslations();
+        $this->translations = $translations;
     }
 
     /**
@@ -96,64 +91,6 @@ class CronExpression
         }
 
         return $this->pluralize($translation);
-    }
-
-    /**
-     * Ensure the locale exists or use a fallback
-     *
-     * @param string $fallbackLocale The fallback locale
-     */
-    protected function ensureLocaleExists(string $fallbackLocale = 'en'): void
-    {
-        if (!is_dir($this->getTranslationDirectory())) {
-            $this->locale = $fallbackLocale;
-        }
-    }
-
-    /**
-     * Load the translation files
-     *
-     * @throws TranslationFileMissingException
-     */
-    protected function loadTranslations(): void
-    {
-        $this->translations = [
-            'days' => $this->loadTranslationFile('days'),
-            'fields' => $this->loadTranslationFile('fields'),
-            'months' => $this->loadTranslationFile('months'),
-            'ordinals' => $this->loadTranslationFile('ordinals'),
-            'times' => $this->loadTranslationFile('times'),
-        ];
-    }
-
-    /**
-     * Load a single translation file
-     *
-     * @param string $file The file name 
-     *
-     * @return array
-     *
-     * @throws TranslationFileMissingException
-     */
-    protected function loadTranslationFile(string $file): array
-    {
-        $filename = sprintf('%s/%s.php', $this->getTranslationDirectory(), $file);
-
-        if (!is_file($filename)) {
-            throw new TranslationFileMissingException($this->locale, $file);
-        }
-
-        return include $filename;
-    }
-
-    /**
-     * Get the translation directory
-     *
-     * @return string
-     */
-    protected function getTranslationDirectory(): string
-    {
-        return __DIR__ . '/lang/' . $this->locale;
     }
 
     /**
